python/trips_recorder/requirements.txt
bcrypt
nanoid
paho-mqtt
peewee
psycopg2-binary
PyMySQL
valkey[hiredis]
resettabletimer